Es 117 & ES-122

+++ src/main/kotlin/org/malachite/estella/people/api/UserController.kt
@@ -1,5 +1,7 @@
 package org.malachite.estella.people.api
 
+import org.malachite.estella.commons.EStellaHeaders
+import org.malachite.estella.commons.OwnResponses
 import org.malachite.estella.commons.Message
 import org.malachite.estella.commons.OneStringValueMessage
 import org.malachite.estella.commons.SuccessMessage
@@ -10,7 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import javax.servlet.http.HttpServletResponse
 
 
 @RestController
@@ -28,7 +29,7 @@ class UserController(
 
     @CrossOrigin
     @PostMapping("/login")
-    fun loginUser(@RequestBody body: LoginRequest, response: HttpServletResponse): ResponseEntity<OneStringValueMessage> {
+    fun loginUser(@RequestBody body: LoginRequest): ResponseEntity<Message> {
         val user = userService.getUserByEmail(body.mail)
             ?: return ResponseEntity(
                 Message("User with such email: ${body.mail} not found"),
@@ -38,36 +39,33 @@ class UserController(
         if (!user.comparePassword(body.password))
             return ResponseEntity(Message("Invalid password"), HttpStatus.BAD_REQUEST)
 
-        val token = securityService.getTokens(user, response)
-        return token?.let { ResponseEntity(Token(token), HttpStatus.OK) }
-            ?: ResponseEntity(Message("Error while creating token"), HttpStatus.INTERNAL_SERVER_ERROR)
+        val tokens = securityService.getTokens(user)
+        return tokens?.let {
+            ResponseEntity.ok()
+                .header(EStellaHeaders.authToken, it.first)
+                .header(EStellaHeaders.refreshToken, it.second)
+                .body(Message("Success"))
+        } ?: ResponseEntity(Message("Error while creating token"), HttpStatus.INTERNAL_SERVER_ERROR)
     }
 
 
     @CrossOrigin
     @GetMapping("/loggedInUser")
-    fun getLoggedInUser(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
-        val user = securityService.getUserFromJWT(jwt)
-        return user?.let {
-            ResponseEntity(user, HttpStatus.OK)
-        } ?: ResponseEntity(Message("Unauthenticated"), HttpStatus.UNAUTHORIZED)
+    fun getLoggedInUser(@RequestHeader(EStellaHeaders.jwtToken) jwt: String?): ResponseEntity<Any> {
+        val user = securityService.getUserFromJWT(jwt) ?: OwnResponses.UNAUTH
+        return ResponseEntity.ok(user)
     }
 
     @CrossOrigin
-    @PostMapping("/logout")
-    fun logout(response: HttpServletResponse): ResponseEntity<Message> {
-        securityService.deleteCookie(response)
-        return ResponseEntity(SuccessMessage, HttpStatus.OK)
-    }
-
-    @CrossOrigin
-    @PostMapping("/refreshToken")
-    fun refresh(@RequestBody token: String,@CookieValue("jwt") jwt: String?,
-                response: HttpServletResponse): ResponseEntity<Message> {
-
-        return securityService.refreshToken(token,jwt, response)
-            ?.let { ResponseEntity.ok(SuccessMessage) }
-            ?: ResponseEntity.status(404).body(Message("Failed during refreshing not found user"))
+    @PostMapping("/{userId}/refreshToken")
+    fun refresh(
+        @PathVariable userId: Int,
+        @RequestHeader(EStellaHeaders.jwtToken) jwt: String?
+    ): ResponseEntity<Message> {
+        jwt ?: return OwnResponses.UNAUTH
+        return securityService.refreshToken(jwt, userId)
+            ?.let { ResponseEntity.ok().header(EStellaHeaders.authToken, it).body(SuccessMessage) }
+            ?: ResponseEntity.status(404).body((Message("Failed during refreshing not found user")))
     }
 
     @CrossOrigin
@@ -86,23 +84,36 @@ class UserController(
 
     @CrossOrigin
     @PutMapping("/{userId}")
-    fun updateUser(@PathVariable("userId") userId: Int, @RequestBody user: UserRequest): ResponseEntity<Message> {
+    fun updateUser(
+        @RequestHeader(EStellaHeaders.jwtToken) jwt: String?,
+        @PathVariable("userId") userId: Int,
+        @RequestBody user: UserRequest
+    ): ResponseEntity<Message> {
+        if (!securityService.checkUserRights(jwt, userId)) return OwnResponses.UNAUTH
         userService.updateUser(userId, user.toUser())
-        return ResponseEntity(SuccessMessage, HttpStatus.OK)
+        return OwnResponses.SUCCESS
     }
 
     @CrossOrigin
     @DeleteMapping("/{userId}")
-    fun deleteUser(@PathVariable("userId") userId: Int): ResponseEntity<Message> {
+    fun deleteUser(
+        @RequestHeader(EStellaHeaders.jwtToken) jwt: String?,
+        @PathVariable("userId") userId: Int
+    ): ResponseEntity<Message> {
+        if (!securityService.checkUserRights(jwt, userId)) return OwnResponses.UNAUTH
         userService.deleteUser(userId)
         return ResponseEntity(SuccessMessage, HttpStatus.OK)
     }
 
 }
 
-data class UserRequest(val firstName: String, val lastName: String, val mail: String, val password: String) {
+data class UserRequest(
+    val firstName: String,
+    val lastName: String,
+    val mail: String,
+    val password: String
+) {
     fun toUser() = User(null, firstName, lastName, mail, password)
 }
 
 data class LoginRequest(val mail: String, val password: String)
-data class Token(val token: String): OneStringValueMessage()
